+++ plugins/npc_control/NPCControl.h
@@ -9,30 +9,30 @@
 namespace Plugins::Npc
 {
 	//! A struct that represents an npc that can be spawned
-	struct Npc : Reflectable
+	struct Npc final : Reflectable
 	{
 		std::string shipArch = "ge_fighter";
 		std::string loadout = "MP_ge_fighter";
 		std::string iff = "fc_fl_grp";
-		uint iffId;
+		uint iffId = 0;
 		uint infocardId = 197808;
 		uint infocard2Id = 197809;
 		std::string pilot = "pilot_pirate_ace";
 		std::string graph = "FIGHTER"; // NOTHING, FIGHTER, TRANSPORT, GUNBOAT, CRUISER. Possibly (unconfirmed) MINER, CAPITAL, FREIGHTER
 
-		uint shipArchId;
-		uint loadoutId;
+		uint shipArchId = 0;
+		uint loadoutId = 0;
 	};
 
 	// A struct that represents a fleet that can be spawned
-	struct Fleet : Reflectable
+	struct Fleet final : Reflectable
 	{
 		std::wstring name = L"example";
 		std::map<std::wstring, int> member = {{L"example", 5}};
 	};
 
 	// A struct that represents an NPC that is spawned on startup
-	struct StartupNpc : Reflectable
+	struct StartupNpc final : Reflectable
 	{
 		std::wstring name = L"example";
 		std::string system = "li01";
@@ -46,7 +46,7 @@ namespace Plugins::Npc
 	};
 
 	//! Config data for this plugin
-	struct Config : Reflectable
+	struct Config final : Reflectable
 	{
 		//! Map of npcs that can be spawned
 		std::map<std::wstring, Npc> npcInfo = {{L"example", Npc()}};
@@ -60,6 +60,16 @@ namespace Plugins::Npc
 		std::string File() override { return "config/npc.json"; }
 	};
 
+	//! Communicator class for this plugin. This is used by other plugins
+	class NpcCommunicator : public PluginCommunicator
+	{
+	  public:
+		inline static const char* pluginName = "NPC Control";
+		explicit NpcCommunicator(const std::string& plug);
+
+		uint PluginCall(CreateNpc, const std::wstring& name, Vector position, const Matrix& rotation, SystemId systemId, bool varyPosition);
+	};
+
 	//! Global data for this plugin
 	struct Global final
 	{
@@ -67,7 +77,8 @@ namespace Plugins::Npc
 		ReturnCode returnCode = ReturnCode::Default;
 		std::vector<const char*> listGraphs {};
 		std::vector<uint> spawnedNpcs {};
-		std::shared_ptr<spdlog::logger> Log = nullptr;
+		std::shared_ptr<spdlog::logger> log = nullptr;
 		uint dockNpc = 0;
+		NpcCommunicator* communicator = nullptr;
 	};
 } // namespace Plugins::Npc